Nother permute problem
git-svn-id: http://svn.code.sf.net/p/coda-oss/code/trunk@153 1dbd42e5-0d20-4692-8756-183d5fc073b5

+++ modules/c++/math.linear/include/math/linear/MatrixMxN.h
@@ -1195,7 +1195,7 @@ template<size_t _ND, typename _T> inline
     for (size_t i = 0; i < _ND; i++)
         a(i, i) = 1;
 
-    int pivots[_ND];
+    size_t pivots[_ND];
     MatrixMxN<_ND, _ND, _T> lu = mx.decomposeLU(pivots);
     
     for (size_t i = 0; i < _ND; i++)
